--- snippets/csharp/VS_Snippets_CLR/formatting.numeric.custom/cs/literal1.cs
@@ -1,22 +1,0 @@
-using System;
-
-public class Example
-{
-   public static void Main()
-   {
-      int n = 9;
-      Console.WriteLine($@"{n:##.0\%}");  
-      Console.WriteLine($@"{n:\'##\'}");      
-      Console.WriteLine($@"{n:\\##\\}");      
-
-      Console.WriteLine($"{n:##.0'%'}");  
-      Console.WriteLine($@"{n:'\'##'\'}");      
-   }
-}
-// The example displays the following output:
-//      9.0%
-//      '9'
-//      \9\
-//
-//      9.0%
-//      \9\
