--- snippets/csharp/VS_Snippets_CLR/formatting.numeric.custom/cs/literal2.cs
@@ -1,12 +1,0 @@
-using System;
-
-public class Example
-{
-   public static void Main()
-   {
-      double n = 123.8;
-      Console.WriteLine($"{n:#,##0.0K}");  
-   }
-}
-// The example displays the following output:
-//      123.8K
