+++ snippets/visualbasic/VS_Snippets_CLR/formatting.numeric.custom/literal2.vb
@@ -1,10 +1,12 @@
 Public Module Example
    Public Sub Main()
+      ' <Snippet1>
       Dim n As Double = 123.8
       Console.WriteLine($"{n:#,##0.0K}")
+      ' The example displays the following output:
+      '       123.8K
+      ' </Snippet1>
     End Sub
 End Module 
-' The example displays the following output:
-'       123.8K
